Fix compilation warning in ANGLETest
BUG=angleproject:1105

Reviewed-on: https://chromium-review.googlesource.com/287831
Tested-by: Corentin Wallez <cwallez@chromium.org>

+++ src/tests/test_utils/ANGLETest.h
@@ -108,8 +108,8 @@ class ANGLETest : public ::testing::TestWithParam<angle::PlatformParameters>
     bool destroyEGLContext();
 
     EGLWindow *mEGLWindow;
-    size_t mWidth;
-    size_t mHeight;
+    int mWidth;
+    int mHeight;
 
     static OSWindow *mOSWindow;
 };
